Epmlstrtai 680 (#75)
Added opportunity to reflect status when import/export stories. (When we import jira stories and if status equals "resolved" we set "PASSED" status to our story (suit in BDD) (conversely when exporting), else we set "NOT_RUN" status. When we export stories from BDD, if we have "FAILED" status and the previous one was "PASSED" (to know the previous state we added versioning of suits), we set "reopened" status. In other cases, the status remains the same as before.